requirements.txt
django
Pillow==2.9.0
celery
redis
#=============================================

#REST Framework

django-phonenumber-field
djangorestframework
pygments
httpie
markdown
django-filter
django-crispy-forms

#=============================================